Matrix op fixes.

+++ jgl-opengl/src/test/java/org/jgl/opengl/test/T002Rect.java
@@ -46,6 +46,7 @@ public class T002Rect extends GL3EventListener {
 
 	@Override
 	protected void onResize(GL3 gl, GlViewSize newViewport) {
-		gl.glViewport(newViewport.x, newViewport.y, newViewport.width, newViewport.height);
+		gl.glViewport(newViewport.x, newViewport.y, 
+				(int) newViewport.width, (int) newViewport.height);
 	}
 }
