+++ jgl-opengl/src/test/java/org/jgl/opengl/test/T005Manderblot.java
@@ -48,6 +48,7 @@ public class T005Manderblot extends GL3EventListener {
 
 	@Override
 	protected void onResize(GL3 gl, GlViewSize newViewport) {
-		gl.glViewport(newViewport.x, newViewport.y, newViewport.width, newViewport.height);
+		gl.glViewport(newViewport.x, newViewport.y, 
+				(int) newViewport.width, (int) newViewport.height);
 	}
 }
